+++ game/cli.py
@@ -1,12 +1,16 @@
 from game.chess import Chess
+from game.board import Board
 
 class RangeError(Exception):
     pass
 
 def main():
+    board = Board()
+    print(board)
     chess = Chess()
     while chess.is_playing():
         play(chess)
+    
 
 def play(chess):
     try:
@@ -37,8 +41,8 @@ def play(chess):
         print('Coords out of range. Please try again')
         
 def validate_input(value):
-    if value < 1 or value > 8:
+    if value < 0 or value > 7:
         raise RangeError
-   
+
 if __name__ == '__main__':
     main()