+++ game/rook.py
@@ -1,8 +1,57 @@
 from game.piece import Piece
 
 class Rook(Piece):
-    def __str__(self):
-        if self.__color__ == "White":
-            return "♜"
-        else:
-            return "♖"
+    def white_str(self):
+        return "♖"
+
+    def black_str(self):
+        return "♜"
+
+    def possible_positions_vd(self, row, col):
+        possibles = []
+        for next_row in range(row + 1, 8):
+            other_piece = self.__board__.get_piece(next_row, col)
+            if other_piece is not None:
+                if other_piece.__color__ != self.__color__:
+                    possibles.append((next_row, col))
+                break
+            possibles.append((next_row, col))
+        return possibles
+
+    def possible_positions_va(self, row, col):
+        possibles = []
+        for next_row in range(row - 1, -1, -1):
+            other_piece = self.__board__.get_piece(next_row, col)
+            if other_piece is not None:
+                if other_piece.__color__ != self.__color__:
+                    possibles.append((next_row, col))
+                break
+            possibles.append((next_row, col))
+        return possibles
+
+    def possible_positions_hl(self, row, col):
+        possibles = []
+        for next_col in range(col + 1, 8):
+            other_piece = self.__board__.get_piece(row, next_col)
+            if other_piece is not None:
+                if other_piece.__color__ != self.__color__:
+                    possibles.append((row, next_col))
+                break
+            possibles.append((row, next_col))
+        return possibles
+    
+    def possible_positions_hr(self, row, col):
+        possibles = []
+        for next_col in range(col - 1, -1, -1):
+            other_piece = self.__board__.get_piece(row, next_col)
+            if other_piece is not None:
+                if other_piece.__color__ != self.__color__:
+                    possibles.append((row, next_col))
+                break
+            possibles.append((row, next_col))
+        return possibles
+    
+
+
+    
+