+++ CHANGELOG.md
@@ -5,6 +5,57 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2024-09-08
+### Fixed
+- Fixed board and cli implementation.
+
+## [0.11.0] - 2024-09-07
+### Added
+- Added piece out of board exception.
+
+## [0.11.0] - 2024-09-06
+### Added
+- Added chess exceptions.
+
+## [0.11.0] - 2024-09-05
+### Added
+- Added get color method in board.
+
+## [0.11.0] - 2024-09-04
+### Added
+- Added move method in board.
+
+## [0.11.0] - 2024-09-03
+### Added
+- Added bishop movements and testing.
+
+## [0.17.0] - 2024-09-02
+### Added
+- Tested piece implementation.
+
+## [0.16.0] - 2024-09-01
+### Added
+- Added tower movements and testing.
+
+## [0.15.0] - 2024-08-29
+### Fixed
+- Fixed board and added pieces in initial position.
+
+### Added
+- Added rook testing
+
+## [0.14.0] - 2024-08-27
+### Added
+- Implemented rook vertical movement
+
+## [0.13.0] - 2024-08-25
+### Fixed
+- Fixed board testing and added place piece method in board file.
+
+## [0.120] - 2024-08-24
+### Fixed
+- Fixed validate_input method in chess file and tested cli.
+
 ## [0.11.0] - 2024-08-23
 ### Added
 - Added is_playing and board_display testing to chess implementation.